+++ NPCs/NoxiumGlobalNPC.cs
@@ -30,11 +30,11 @@ namespace NoxiumMod.NPCs
 					}
 					Main.NewText("You sense visionary essence from below", 70, 140, 80, false);
 
-					for (int j = 0; j < (int)(Main.rockLayer * (double)Main.maxTilesY * 0.0015); j++) //the 0.0015 number basically changes how much of the ore spawns
+					for (int j = 0; j < (int)(Main.rockLayer * Main.maxTilesY * 0.0015); j++) //the 0.0015 number basically changes how much of the ore spawns
 					{
 						int num3 = Main.rand.Next(0, Main.maxTilesX);
 						int num4 = Main.rand.Next((int)Main.rockLayer, Main.maxTilesY - 200);
-						WorldGen.OreRunner(num3, num4, (double)Main.rand.Next(5, 6), Main.rand.Next(7, 8), (ushort)mod.TileType("OculumOre"));
+						WorldGen.OreRunner(num3, num4, Main.rand.Next(5, 6), Main.rand.Next(7, 8), (ushort)mod.TileType("OculumOre"));
 					}
 				}
 				NoxiumWorld.oculumOreSpawn = true;
