+++ Projectiles/CrystalCannonHitProj.cs
@@ -44,7 +44,7 @@ namespace NoxiumMod.Projectiles
 			this.Timer--;
 			if (this.Timer <= 10)
 			{
-				projectile.rotation = (float)Math.Atan2((double)projectile.velocity.Y, (double)projectile.velocity.X) + 1.57f;
+				projectile.rotation = (float)Math.Atan2(projectile.velocity.Y, projectile.velocity.X) + 1.57f;
 
 				float num2 = projectile.Center.X;
 				float num3 = projectile.Center.Y;
@@ -54,9 +54,9 @@ namespace NoxiumMod.Projectiles
 				{
 					if (Main.npc[j].CanBeChasedBy(projectile, false) && Collision.CanHit(projectile.Center, 1, 1, Main.npc[j].Center, 1, 1))
 					{
-						float num5 = Main.npc[j].position.X + (float)(Main.npc[j].width / 2);
-						float num6 = Main.npc[j].position.Y + (float)(Main.npc[j].height / 2);
-						float num7 = Math.Abs(projectile.position.X + (float)(projectile.width / 2) - num5) + Math.Abs(projectile.position.Y + (float)(projectile.height / 2) - num6);
+						float num5 = Main.npc[j].position.X + Main.npc[j].width / 2;
+						float num6 = Main.npc[j].position.Y + Main.npc[j].height / 2;
+						float num7 = Math.Abs(projectile.position.X + projectile.width / 2 - num5) + Math.Abs(projectile.position.Y + projectile.height / 2 - num6);
 						if (num7 < num4)
 						{
 							num4 = num7;
@@ -69,10 +69,10 @@ namespace NoxiumMod.Projectiles
 				if (flag)
 				{
 					float num8 = 26f;
-					Vector2 vector = new Vector2(projectile.position.X + (float)projectile.width * 0.5f, projectile.position.Y + (float)projectile.height * 0.5f);
+					Vector2 vector = new Vector2(projectile.position.X + projectile.width * 0.5f, projectile.position.Y + projectile.height * 0.5f);
 					float num9 = num2 - vector.X;
 					float num10 = num3 - vector.Y;
-					float num11 = (float)Math.Sqrt((double)(num9 * num9 + num10 * num10));
+					float num11 = (float)Math.Sqrt(num9 * num9 + num10 * num10);
 					num11 = num8 / num11;
 					num9 *= num11;
 					num10 *= num11;
@@ -89,7 +89,7 @@ namespace NoxiumMod.Projectiles
 			}
 			else
 			{
-				projectile.rotation += 0.2f * (float)projectile.direction;
+				projectile.rotation += 0.2f * projectile.direction;
 				projectile.ai[0] += 1f;
 				projectile.localAI[0] += 1f;
 			}
