+++ Projectiles/PeriwumProjectile.cs
@@ -44,8 +44,8 @@ namespace NoxiumMod.Projectiles
 			projectile.ai[1] += 4f; // How fast it circles the player
 			for (int i = 0; i < 200; i++)
 			{
-				float shootToX = Main.npc[i].position.X + (float)Main.npc[i].width * 0.5f - projectile.Center.X;
-				float shootToY = Main.npc[i].position.Y + (float)Main.npc[i].height * 0.5f - projectile.Center.Y;
+				float shootToX = Main.npc[i].position.X + Main.npc[i].width * 0.5f - projectile.Center.X;
+				float shootToY = Main.npc[i].position.Y + Main.npc[i].height * 0.5f - projectile.Center.Y;
 				Projectile.NewProjectile(projectile.position.X, projectile.position.Y, shootToX, shootToY, ModContent.ProjectileType<PeriwumProjectile_Small>(), 31, projectile.knockBack, Main.myPlayer);
 
 			}
